// src/types/bangumi_api.ts
export interface BangumiComponents {
  schemas: {
// ...
    User1: {
      /** ID */
      id: number;
      /** URL */
      url: string;
      /**
       * Username
       * @description 唯一用户名，初始与 UID 相同，可修改一次
       */
      username: string;
      /** Nickname */
      nickname: string;
      user_group: BangumiComponents["schemas"]["UserGroup"];
      avatar: BangumiComponents["schemas"]["Avatar"];
      /**
       * Sign
       * @description 个人签名
       */
      sign: string;
    };
// ...
    Paged_UserCollection_: {
      /** Total */
      total?: number;
      /** Limit */
      limit?: number;
      /** Offset */
      offset?: number;
      /**
       * Data
       * @default
       */
      data?: BangumiComponents["schemas"]["UserCollection"][];
    };
// ...
    Subject: {
      /** Id */
      id: number;
      /** Type */
      type: number;
      /** Name */
      name: string;
      /** Name Cn */
      name_cn: string;
      /** Summary */
      summary: string;
      /** Nsfw */
      nsfw: boolean;
      /** Locked */
      locked: boolean;
      /**
       * Date
       * @description air date in `YYYY-MM-DD` format
       */
      date?: string;
      /**
       * Platform
       * @description TV, Web, 欧美剧, PS4...
       */
      platform: string;
      images?: BangumiComponents["schemas"]["Images"];
      /** Infobox */
      infobox?: BangumiComponents["schemas"]["Item"][];
      /**
       * Volumes
       * @description 书籍条目的册数，由旧服务端从wiki中解析
       */
      volumes: number;
      /**
       * Eps
       * @description 由旧服务端从wiki中解析，对于书籍条目为`话数`
       */
      eps: number;
      /**
       * Total Episodes
       * @description 数据库中的章节数量
       */
      total_episodes: number;
      rating: BangumiComponents["schemas"]["Rating"];
      collection: BangumiComponents["schemas"]["Collection"];
      /** Tags */
      tags: BangumiComponents["schemas"]["Tag"][];
    };
    /**
     * SubjectType
     * @description 条目类型
     * - `1` 为 书籍
     * - `2` 为 动画
     * - `3` 为 音乐
     * - `4` 为 游戏
     * - `6` 为 三次元
     *
     * 没有 `5`
     * @enum {number}
     */
    SubjectType: 1 | 2 | 3 | 4 | 6;
    /** Tag */
    Tag: {
      /** Name */
      name: string;
      /** Count */
      count: number;
    };
    /** UserCollection */
    UserCollection: {
      /** Subject Id */
      subject_id: number;
      subject_type: BangumiComponents["schemas"]["SubjectType"];
      /** Rate */
      rate: number;
      type: BangumiComponents["schemas"]["CollectionType"];
      /** Comment */
      comment?: string;
      /** Tags */
      tags: string[];
      /** Ep Status */
      ep_status: number;
      /** Vol Status */
      vol_status: number;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
      /** Private */
      private: boolean;
    };
// ...
  };
// ...
  parameters: {
    /** @description 用户名 <br> 也可使用 UID */
    username: string;
    /** @description 条目 ID */
    subject_id: number;
    /** @description 章节 ID */
    ep_id: number;
    /** @description 收视类型，参考 [EpStatusType](#model-EpStatusType) */
    ep_status: BangumiComponents["schemas"]["EpStatusType"];
    /** @description 返回数据大小，参考 [ResponseGroup](#model-ResponseGroup) <br> 默认为 small */
    responseGroup: BangumiComponents["schemas"]["ResponseGroup"];
    /** @description [https://bgm.tv/dev/app](https://bgm.tv/dev/app) 申请到的 App ID */
    app_id: string;
  };
}

export interface BangumiOperations {
// ...
  getUserCollectionsByUsername: {
    parameters: {
      path: {
        /** 设置了 username 后无法使用UID */
        username: string;
      };
      query: {
        /**
         * 条目类型，默认为全部
         *
         * 具体含义见 [SubjectType](#model-SubjectType)
         */
        subject_type?: number;
        /**
         * 收藏类型，默认为全部
         *
         * 具体含义见 [CollectionType](#model-CollectionType)
         */
        type?: number;
        limit?: number;
        offset?: number;
      };
    };
    responses: {
      /** Successful Response */
      200: {
        content: {
          "application/json": BangumiComponents["schemas"]["Paged_UserCollection_"];
        };
      };
      /** Validation Error */
      400: {
        content: {
          "application/json": BangumiComponents["schemas"]["ErrorDetail"];
        };
      };
      /** 用户不存在 */
      404: {
        content: {
          "application/json": BangumiComponents["schemas"]["ErrorDetail"];
        };
      };
    };
  };
// ...
}